+++ crates/objc2/tests/declare_class.rs
@@ -2,7 +2,7 @@
 use core::ptr::{self, NonNull};
 
 use objc2::declare::IvarEncode;
-use objc2::mutability::{Immutable, Mutable};
+use objc2::mutability::Immutable;
 use objc2::rc::Id;
 use objc2::runtime::NSObject;
 use objc2::{declare_class, extern_methods, sel, ClassType};
@@ -386,71 +386,6 @@ fn test_duplicate_ivar() {
     );
 
     let _ = DeclareClassDuplicateIvar::class();
-}
-
-#[test]
-#[should_panic = "instance variable \"ivar\" already exists on a superclass"]
-fn test_subclass_duplicate_ivar() {
-    declare_class!(
-        struct Cls {
-            ivar_superclass: IvarEncode<i32, "ivar">,
-        }
-
-        mod ivars;
-
-        unsafe impl ClassType for Cls {
-            type Super = NSObject;
-            type Mutability = Mutable;
-            const NAME: &'static str = "DeclareClassDuplicateIvarSuperclass";
-        }
-    );
-
-    declare_class!(
-        struct SubCls {
-            ivar_subclass: IvarEncode<i32, "ivar">,
-        }
-
-        mod ivars_subclass;
-
-        unsafe impl ClassType for SubCls {
-            type Super = Cls;
-            type Mutability = Mutable;
-            const NAME: &'static str = "DeclareClassDuplicateIvarSubclass";
-        }
-    );
-
-    extern_methods!(
-        unsafe impl SubCls {
-            #[method_id(new)]
-            fn new() -> Id<Self>;
-        }
-    );
-
-    let _ = SubCls::class();
-
-    // The rest is to show what would go wrong if it didn't panic
-
-    assert_eq!(Cls::class().instance_size(), 16);
-    assert_eq!(SubCls::class().instance_size(), 16);
-
-    let mut obj = SubCls::new();
-
-    // Zero-initialized
-    assert_eq!(*obj.ivar_superclass, 0);
-    assert_eq!(*obj.ivar_subclass, 0);
-
-    *obj.ivar_subclass = 2;
-
-    assert_eq!(*obj.ivar_superclass, 2);
-    assert_eq!(*obj.ivar_subclass, 2);
-
-    *obj.ivar_superclass = 3;
-
-    assert_eq!(*obj.ivar_superclass, 3);
-    assert_eq!(*obj.ivar_subclass, 3);
-
-    let ivar_dynamically = unsafe { obj.ivar::<i32>("ivar") };
-    assert_eq!(*ivar_dynamically, 3);
 }
 
 declare_class!(